+++ archetype-pom/java-project/src/main/resources/META-INF/archetype-post-generate.groovy
@@ -36,13 +36,19 @@ def removeRootProjectLaunchers() {
 def removeRootProjectPomConfiguration() {
     projectPom = projectDir.resolve("pom.xml")
     out = Files.readAllLines(projectPom)
+    // Remove SCM configuration
+    for (int i = 96; i >= 92; i--) {
+        out.remove(i);
+    }
     // Remove maven-site-plugin configuration
-    for (int i = 94; i >= 88; i--) {
+    for (int i = 87; i >= 81; i--) {
         out.remove(i);
     }
     // Remove Sonar project key configuration
-    out.remove(45);
-    out.remove(44);
+    out.remove(38);
+    out.remove(37);
+    // Remove project URL configuration
+    out.remove(20);
     Files.write(projectPom, out, StandardOpenOption.WRITE, StandardOpenOption.TRUNCATE_EXISTING)
 }